Split commands/publish into focused submodules
publish/mod.rs keeps the create/sync entry points and shared tests;
scope.rs owns repo selection, provider filtering, and base overrides;
remote.rs owns per-repo publish execution (workspace and artifact);
sync.rs owns PR body cross-link sync; status.rs owns the status views.
Pure moves, no behavior change.

Knit-Group: kg_20260611_47b60a
Knit-Bundle: open-source-prep
